Validação dos dados de perfil de usuário

+++ app/Models/User.php
@@ -58,13 +58,17 @@ class User extends Authenticatable implements JWTSubject
     public function updateWithProfile(array $data): User 
     {
         try {
-            UserProfile::uploadPhoto($data['photo']);
+            if (isset($data['photo'])) {
+                UserProfile::uploadPhoto($data['photo']);
+            }
             \DB::beginTransaction();
             $this->fill($data)->save();
             UserProfile::saveProfile($this, $data);
             \DB::commit();
         } catch (Exception $e) {
-            UserProfile::deleteFile($data);
+            if (isset($data['photo'])) {
+                UserProfile::deleteFile($data);
+            }
             \DB::rollBack();
             throw $e;
         }
