+++ app/Models/UserProfile.php
@@ -21,8 +21,10 @@ class UserProfile extends Model
 
     public static function saveProfile(User $user, array $data): UserProfile
     {
-        self::deletePhoto($user);
-        $data['photo'] = self::getPhotoHashName($data['photo']);
+        if (array_key_exists('photo', $data)) {
+            self::deletePhoto($user);
+            $data['photo'] = self::getPhotoHashName($data['photo']);
+        }
         $user->profile->fill($data)->save();
         return $user->profile;
     }
